text-mining/requirements.txt
torch
transformers
lancedb
fitz
PyMuPDF
PyPDF2
python-docx
pandas
sentence_transformers
pyarrow
pathlib
langchain
accelerate